+++ quacc/schemas/cclib.py
@@ -10,6 +10,7 @@ from atomate2.common.schemas.cclib import TaskDocument
 
 from quacc.schemas.atoms import atoms_to_metadata
 from quacc.util.atoms import prep_next_run as prep_next_run_
+from quacc.util.dicts import remove_dict_empties
 
 
 def summarize_run(
@@ -19,6 +20,7 @@ def summarize_run(
     check_convergence: bool = True,
     transition_state: bool = False,
     prep_next_run: bool = True,
+    remove_empties: bool = False,
     additional_fields: dict = None,
 ) -> dict:
     """
@@ -47,6 +49,8 @@ def summarize_run(
     prep_next_run
         Whether the Atoms object storeed in {"atoms": atoms} should be prepared for the next run.
         This clears out any attached calculator and moves the final magmoms to the initial magmoms.
+    remove_empties
+        Whether to remove None values and empty lists/dicts from the TaskDocument.
     additional_fields
         Additional fields to add to the task document.
 
@@ -54,50 +58,43 @@ def summarize_run(
     -------
     Dictionary representation of the task document with the following fields:
         - atoms: Atoms = Field(None, title = "The Atoms object obtained from atoms.")
-        - parameters: dict = Field(None, title = "The input parameters obtained from atoms.calc.parameters.")
-        - nid: str = Field(None, title = "The node ID representing the machine where the calculation was run.")
         - atoms_info: dict = Field(None, title = "The Atoms object info obtained from atoms.info.")
-
-        The task document also inherits the following fields from atomate2.common.schemas.cclib.TaskDocument:
-        - molecule: Molecule = Field(None, description="Final output molecule from the task")
-        - energy: float = Field(None, description="Final total energy")
-        - dir_name: str = Field(None, description="Directory where the output is parsed")
-        - logfile: str = Field(None, description="Path to the log file used in the post-processing analysis")
         - attributes: Dict = Field(None, description="Computed properties and calculation outputs. See https://cclib.github.io/data_notes.html")
-        - metadata: Dict = Field(None, description="Calculation metadata, including input parameters and runtime statistics. See https://cclib.github.io/data_notes.html#metadata")
-        - task_label: str = Field(None, description="A description of the task")
-        - tags: List[str] = Field(None, description="Optional tags for this task document")
-        - last_updated: str = Field(default_factory=datetime_str, description="Timestamp for this task document was last updated")
-        - _schema: str = Field(__version__, description="Version of atomate2 used to create the document", alias="schema")
-
-        ... as well as the following fields from emmet.core.structure.MoleculeMetadata:
-        - charge: int = Field(None, description="Charge of the molecule")
-        - spin_multiplicity: int = Field(None, description="Spin multiplicity of the molecule")
-        - natoms: int = Field(None, description="Total number of atoms in the molecule")
-        - elements: List[Element] = Field(None, description="List of elements in the molecule")
-        - nelements: int = Field(None, title="Number of Elements")
-        - nelectrons: int = Field(None, title="Number of electrons", description="The total number of electrons for the molecule")
-        - composition: Composition = Field(None, description="Full composition for the molecule")
-        - composition_reduced: Composition = Field(None, title="Reduced Composition", description="Simplified representation of the composition")
-        - formula_alphabetical: str = Field(None, title="Alphabetical Formula", description="Alphabetical molecular formula")
-        - formula_pretty: str = Field(None, title="Pretty Formula", description="Cleaned representation of the formula.")
-        - formula_anonymous: str = Field(None, title="Anonymous Formula", description="Anonymized representation of the formula")
-        - chemsys: str = Field(None, title="Chemical System", description="dash-delimited string of elements in the molecule")
-        - symmetry: PointGroupData = Field(None, description="Symmetry data for this molecule")
-            - point_group: str = Field(None, title="Point Group Symbol", description="The point group for the lattice")
-            - rotation_number: float = Field(None, title="Rotational Symmetry Number", description="Rotational symmetry number for the molecule")
-            - linear: bool = Field(None, title="Molecule Linearity", description="Is the molecule linear?")
-            - tolerance: float = Field(None, title="Point Group Analyzer Tolerance", description="Distance tolerance to consider sites as symmetrically equivalent.")
-            - eigen_tolerance: float = Field(None, title="Interia Tensor Eigenvalue Tolerance", description="Tolerance to compare eigen values of the inertia tensor.")
-            - matrix_tolerance: float = Field(None, title="Symmetry Operation Matrix Element Tolerance" description="Tolerance used to generate the full set of symmetry operations of the point group.")
-
-        ... as well as the following fields from emmet.core.base.EmmetBaseModel:
         - builder_meta: EmmetMeta = Field(default_factory=EmmetMeta, description="Builder metadata."):
+            - build_date: str = Field(default_factory=datetime.utcnow, description="The build date for this document.")
+            - database_version: str = Field(None, description="The database version for the built data.")
             - emmet_version: str = Field(__version__, description="The version of emmet this document was built with.")
             - pymatgen_version: str = Field(pmg_version, description="The version of pymatgen this document was built with.")
             - pull_request: int = Field(None, description="The pull request number associated with this data build.")
-            - database_version: str = Field(None, description="The database version for the built data.")
-            - build_date: str = Field(default_factory=datetime.utcnow, description="The build date for this document.")
+        - charge: int = Field(None, description="Charge of the molecule")
+        - chemsys: str = Field(None, title="Chemical System", description="dash-delimited string of elements in the molecule")
+        - composition: Composition = Field(None, description="Full composition for the molecule")
+        - composition_reduced: Composition = Field(None, title="Reduced Composition", description="Simplified representation of the composition")
+        - dir_name: str = Field(None, description="Directory where the output is parsed")
+        - elements: List[Element] = Field(None, description="List of elements in the molecule")
+        - energy: float = Field(None, description="Final total energy")
+        - formula_alphabetical: str = Field(None, title="Alphabetical Formula", description="Alphabetical molecular formula")
+        - formula_anonymous: str = Field(None, title="Anonymous Formula", description="Anonymized representation of the formula")
+        - formula_pretty: str = Field(None, title="Pretty Formula", description="Cleaned representation of the formula.")
+        - last_updated: str = Field(default_factory=datetime_str, description="Timestamp for this task document was last updated")
+        - logfile: str = Field(None, description="Path to the log file used in the post-processing analysis")
+        - metadata: Dict = Field(None, description="Calculation metadata, including input parameters and runtime statistics. See https://cclib.github.io/data_notes.html#metadata")
+        - molecule: Molecule = Field(None, description="Final output molecule from the task")
+        - natoms: int = Field(None, description="Total number of atoms in the molecule")
+        - nid: str = Field(None, title = "The node ID representing the machine where the calculation was run.")
+        - nelectrons: int = Field(None, title="Number of electrons", description="The total number of electrons for the molecule")
+        - nelements: int = Field(None, title="Number of Elements")
+        - parameters: dict = Field(None, title = "The input parameters obtained from atoms.calc.parameters.")
+        - spin_multiplicity: int = Field(None, description="Spin multiplicity of the molecule")
+        - symmetry: PointGroupData = Field(None, description="Symmetry data for this molecule")
+            - eigen_tolerance: float = Field(None, title="Interia Tensor Eigenvalue Tolerance", description="Tolerance to compare eigen values of the inertia tensor.")
+            - linear: bool = Field(None, title="Molecule Linearity", description="Is the molecule linear?")
+            - matrix_tolerance: float = Field(None, title="Symmetry Operation Matrix Element Tolerance" description="Tolerance used to generate the full set of symmetry operations of the point group.")
+            - point_group: str = Field(None, title="Point Group Symbol", description="The point group for the lattice")
+            - rotation_number: float = Field(None, title="Rotational Symmetry Number", description="Rotational symmetry number for the molecule")
+            - tolerance: float = Field(None, title="Point Group Analyzer Tolerance", description="Distance tolerance to consider sites as symmetrically equivalent.")
+        - tags: List[str] = Field(None, description="Optional tags for this task document")
+        - task_label: str = Field(None, description="A description of the task")
     """
     # Make sure there is a calculator with results
     if not atoms.calc:
@@ -143,11 +140,16 @@ def summarize_run(
     if prep_next_run:
         atoms = prep_next_run_(atoms)
 
-    # Add atoms info
+    # We use get_metadata=False and store_pmg=False because the TaskDocument already
+    # makes the structure metadata for us
     atoms_db = atoms_to_metadata(atoms, get_metadata=False, store_pmg=False)
 
     # Create a dictionary of the inputs/outputs
     task_doc = {**atoms_db, **inputs, **results, **additional_fields}
+
+    # Remove empty fields
+    if remove_empties:
+        task_doc = remove_dict_empties(task_doc)
 
     # Sort dict
     task_doc = dict(sorted(task_doc.items()))